Renamed vars and made fps Uint32 and fixed next_sec

+++ src/something_game.cpp
@@ -452,7 +452,7 @@ void Game::spawn_projectile(Vec2f pos, Vec2f vel, Entity_Index shooter)
     }
 }
 
-void Game::render_debug_overlay(SDL_Renderer *renderer, float fps)
+void Game::render_debug_overlay(SDL_Renderer *renderer, Uint32 fps)
 {
     sec(SDL_SetRenderDrawColor(renderer, 255, 0, 0, 255));
 
@@ -471,7 +471,7 @@ void Game::render_debug_overlay(SDL_Renderer *renderer, float fps)
              FONT_DEBUG_COLOR,
              FONT_SHADOW_COLOR,
              vec2(PADDING, PADDING),
-             "FPS: %.0f", fps);
+             "FPS: %.0d", fps);
     displayf(renderer, &debug_font,
              FONT_DEBUG_COLOR,
              FONT_SHADOW_COLOR,
